+++ crates/infra/curvine-sys/src/cache_manager.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::sys;
-use crate::sys::{CInt, SysResult};
+use crate as sys;
+use crate::{CInt, SysResult};
 use std::fs::File;
 
 const LONG_READ_THRESHOLD_LEN: i64 = 256 * 1024;
